각종 오류 수정

+++ src/main/java/com/moeum/moeum/api/ledger/investSummary/dto/InvestSummaryResponseDto.java
@@ -2,7 +2,7 @@ package com.moeum.moeum.api.ledger.investSummary.dto;
 
 public record InvestSummaryResponseDto(
         Long id,
-        Long year,
-        Long month,
+        Integer year,
+        Integer month,
         Long principal
 ) {}
